剛剛多送了小K的測試版 ClientThread.java 少了BAO提供的一個檔 補上 修改poolsize
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1042 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/datatables/ItemTable.java
@@ -83,7 +83,7 @@ public class ItemTable {
 		_etcItemTypes.put("other", new Integer(12));
 		_etcItemTypes.put("material", new Integer(13));
 		_etcItemTypes.put("event", new Integer(14));
-		_etcItemTypes.put("sting", new Integer(15));
+		_etcItemTypes.put("sting", new Integer(20));
 		_etcItemTypes.put("treasure_box", new Integer(16));
 
 		_useTypes.put("none", new Integer(-1)); // 使用不可能
